encoding: Split out macros from maccommands
