+++ lorawan-encoding/src/lib.rs
@@ -9,6 +9,7 @@ pub mod creator;
 pub mod keys;
 pub mod maccommandcreator;
 pub mod maccommands;
+mod macros;
 pub mod packet_length;
 pub mod parser;
 pub mod string;
